f refactor DbClient in endpoint
