+++ autoendpoint/src/extractors/routers.rs
@@ -71,7 +71,7 @@ impl FromRequest for Routers {
 
         future::ok(Routers {
             webpush: WebPushRouter {
-                ddb: state.ddb.clone(),
+                ddb: state.db_client.clone(),
                 metrics: state.metrics.clone(),
                 http: state.http.clone(),
                 endpoint_url: state.settings.endpoint_url(),
