+++ autoendpoint/src/routers/adm/router.rs
@@ -174,7 +174,7 @@ mod tests {
     use crate::routers::common::tests::{make_notification, CHANNEL_ID};
     use crate::routers::RouterError;
     use crate::routers::{Router, RouterResponse};
-    use autopush_common::db::dynamodb::UserRecord;
+    use autopush_common::db::UserRecord;
     use cadence::StatsdClient;
     use mockall::predicate;
     use serde_json::Value;
