+++ autoendpoint/src/routes/health.rs
@@ -11,8 +11,8 @@ use std::thread;
 
 /// Handle the `/health` and `/__heartbeat__` routes
 pub async fn health_route(state: Data<ServerState>) -> Json<serde_json::Value> {
-    let router_health = interpret_table_health(state.ddb.router_table_exists().await);
-    let message_health = interpret_table_health(state.ddb.message_table_exists().await);
+    let router_health = interpret_table_health(state.db_client.router_table_exists().await);
+    let message_health = interpret_table_health(state.db_client.message_table_exists().await);
 
     Json(json!({
         "status": "OK",
