+++ autoendpoint/src/routes/registration.rs
@@ -9,7 +9,7 @@ use crate::headers::util::get_header;
 use crate::server::ServerState;
 use actix_web::web::{Data, Json};
 use actix_web::{HttpRequest, HttpResponse};
-use autopush_common::db::DynamoDbUser;
+use autopush_common::db::UserRecord;
 use autopush_common::endpoint::make_endpoint;
 use cadence::{CountedExt, StatsdClient};
 use uuid::Uuid;
@@ -33,18 +33,18 @@ pub async fn register_uaid_route(
     incr_metric("ua.command.register", &state.metrics, &request);
 
     // Register user and channel in database
-    let user = DynamoDbUser {
+    let user = UserRecord {
         router_type: path_args.router_type.to_string(),
         router_data: Some(router_data),
-        current_month: Some(state.ddb.message_table().to_string()),
+        current_month: Some(state.db_client.message_table().to_string()),
         ..Default::default()
     };
     let channel_id = router_data_input.channel_id.unwrap_or_else(Uuid::new_v4);
     trace!("Creating user with UAID {}", user.uaid);
     trace!("user = {:?}", user);
     trace!("channel_id = {}", channel_id);
-    state.ddb.add_user(&user).await?;
-    state.ddb.add_channel(user.uaid, channel_id).await?;
+    state.db_client.add_user(&user).await?;
+    state.db_client.add_channel(user.uaid, channel_id).await?;
 
     // Make the endpoint URL
     trace!("Creating endpoint for user");
@@ -83,7 +83,7 @@ pub async fn unregister_user_route(
     state: Data<ServerState>,
 ) -> ApiResult<HttpResponse> {
     debug!("Unregistering UAID {}", path_args.uaid);
-    state.ddb.remove_user(path_args.uaid).await?;
+    state.db_client.remove_user(path_args.uaid).await?;
     Ok(HttpResponse::Ok().finish())
 }
 
@@ -105,7 +105,7 @@ pub async fn update_token_route(
     let router_data = router.register(&router_data_input, &path_args.app_id)?;
 
     // Update the user in the database
-    let user = DynamoDbUser {
+    let user = UserRecord {
         uaid: path_args.uaid,
         router_type: path_args.router_type.to_string(),
         router_data: Some(router_data),
@@ -113,7 +113,7 @@ pub async fn update_token_route(
     };
     trace!("Updating user with UAID {}", user.uaid);
     trace!("user = {:?}", user);
-    state.ddb.update_user(&user).await?;
+    state.db_client.update_user(&user).await?;
 
     trace!("Finished updating token for UAID {}", user.uaid);
     Ok(HttpResponse::Ok().finish())
@@ -131,7 +131,7 @@ pub async fn new_channel_route(
     let channel_data = channel_data.map(Json::into_inner).unwrap_or_default();
     let channel_id = channel_data.channel_id.unwrap_or_else(Uuid::new_v4);
     trace!("channel_id = {}", channel_id);
-    state.ddb.add_channel(path_args.uaid, channel_id).await?;
+    state.db_client.add_channel(path_args.uaid, channel_id).await?;
 
     // Make the endpoint URL
     trace!("Creating endpoint for the new channel");
@@ -158,7 +158,7 @@ pub async fn get_channels_route(
     state: Data<ServerState>,
 ) -> ApiResult<HttpResponse> {
     debug!("Getting channel IDs for UAID {}", path_args.uaid);
-    let channel_ids = state.ddb.get_channels(path_args.uaid).await?;
+    let channel_ids = state.db_client.get_channels(path_args.uaid).await?;
 
     Ok(HttpResponse::Ok().json(serde_json::json!({
         "uaid": path_args.uaid,
@@ -186,7 +186,7 @@ pub async fn unregister_channel_route(
     );
 
     incr_metric("ua.command.unregister", &state.metrics, &request);
-    let channel_did_exist = state.ddb.remove_channel(path_args.uaid, channel_id).await?;
+    let channel_did_exist = state.db_client.remove_channel(path_args.uaid, channel_id).await?;
 
     if channel_did_exist {
         Ok(HttpResponse::Ok().finish())
