+++ autoendpoint/src/routes/webpush.rs
@@ -27,7 +27,7 @@ pub async fn delete_notification_route(
     debug!("Deleting notification with sort-key {}", sort_key);
     trace!("message_id = {:?}", message_id);
     state
-        .ddb
+        .db_client
         .remove_message(message_id.uaid(), sort_key)
         .await?;
